Limit a single HPSS pull script to ten files per hour.
There can be significan fixed, per-file overheads associated with
pulling files out of HPSS, regardless of the size of the file, due
to latencies associated with fetching particular tapes.

This limits a single four-hour HPSS pull script to 40 files, or
six minutes per file, which seems to be reasonable.  Pushes are
not limited in this way: HPSS injest bandwidth is significantly
higher.