Use "lfs quota" to get cedar_online free space
This shoe-horns-in a invocation of "lfs quota" to check the
actual amount of space available on cedar /project.  Only done for
cedar_online just to reduce the number of times the same quota is
fetched.  (The call can take a few seconds to complete).

Used to automatically adjust the cedar doomsday clock in Grafana.