fix(alpenhornd): Several minor cluster-alpenhornd improvements (#15)
* update bbcp invocation:

* remove period progress output "-P 15", which was some voodoo added
  as a guess that the cause of failed bbcp transfers was due to the
  process zombifying
* remove output ordering "-o"
* change checksum calculation to not need ouput ordering:
  - replace "-E md5=" with "-E %md5=" to calculate the checksum on
    the source, not the destination
  - add block-level checksums to re-introduce network transport check
    "-e"

* reduce max allowed queued slurm jobs from 2 to 1

Because alpenhorn has no memory, when the archivelong HPSS queue on
niagara gets full and pull jobs take multiple hours to complete, then
pulls for the same files tend to get enqueued multiple times.  Reducing
the number of pending jobs should reduce the frequency at which this
happens without significantly impacting throughput during times of no
queue congestion

* Use "lfs quota" to get cedar_online free space

This shoe-horns-in a invocation of "lfs quota" to check the
actual amount of space available on cedar /project.  Only done for
cedar_online just to reduce the number of times the same quota is
fetched.  (The call can take a few seconds to complete).

Used to automatically adjust the cedar doomsday clock in Grafana.

* fix(update): Check existence before absence

Without this fix there's a race condition involving niagara where
we see:

* alpenhorn_transfers cron job on jingle syncs scinet_staging
  to scinet_hpss
* alpenhornd on niagara handles HPSS callback to register a successful
  file transfer to HPSS.  When the archivelong queue is full, this
  can happen hours later from the above due to pushing to HPSS being
  paused.
* alpenhorn_transfers cron job on jingle cleans scinet_staging of
  files present in HPSS
* when the archivelong queue empties out and alpenhornd on niagara
  starts thinking about HPSS pushes again, it realises that one of
  the files it wants to sync (the one that was successfully completed
  and removed from staging) no longer exists, so it just complains
  about it not being present on staging anymore.
* The fix makes it realise first that it has nothing to complain
  about because it wouldn't have transferred it since it's already
  in HPSS.

* Handle non-numeric things in "lfs quota" output.

There will be a trailing * on the usage number when over quota

* black

* fix(update): Don't check source when not needed

When a copy fails, and it's clear that the source file isn't the
cause of the failure, then don't mark the source file suspect because
checking it isn't going to do anything other than waste cycles

* Don't check source if rsync dest couldn't be created

If rsync fails and sterr contains "mkstemp" then the rsync failed to
create the destination file, so don't cause a recheck of the source.

* logging