fix(scripts): Make stopping alpenhorn more robust on robots
This is a change I made to the robot script deployed on cedar during the
last nearline I/O trouble.  The alpenhornd processes were getting stuck
in I/O wait and not getting killed, leading to them piling up over
time as the watchdog restarted the laggy system.

In I/O wait, these processes probably won't respond to SIGKILL, but
at least this will _try_ to keep the alpnehorn processes on the robot
node from getting out-of-hand.

I've made the change to the niagara script, too, but not deployed it.
Because all the HPSS I/O happens in a job, there's less risk of
alpenhornd getting stuck (because the queue manager will just kill it
once the job goes over time).  Nevertheless, I don't see any reason
not to update the niagara script, too.